backend/requirements.txt
fastapi==0.95.2
pydantic==1.10.24
starlette==0.27.0
uvicorn==0.39.0
python-dotenv==1.2.1
certifi==2026.2.25